Added a loop that keeps sending Hello packets to the gamecoordinator, fixes us not being welcomed.

+++ Steamkit2-csgo/Match.cs
@@ -145,7 +145,7 @@ namespace SteamKit.CSGO
         /// <summary>
         /// ID of match
         /// </summary>
-        public uint? MatchId;
+        public ulong? MatchId;
         /// <summary>
         /// ID of outcome of match
         /// </summary>
